Courses Ads 2

+++ routes/courses_ads.php
@@ -7,7 +7,7 @@ use App\Http\Controllers\{
 use Illuminate\Support\Facades\Route;
 
 Route::prefix('courses-ads')->controller(CoursesAdsController::class)->group(function() {
-    Route::post('admin', 'store');
+    Route::post('admin', 'store')->middleware('admin');
     Route::get('', 'index');
-    Route::delete('', 'destroy');
+    Route::delete('admin/{courses_ad}', 'destroy')->middleware('admin');
 });
